RRO: Don't update overlays when target changes
When packages or components are enabled/disabled, PackageManager
broadcasts ACTION_PACKAGE_CHANGED. Currently, overlay paths are being
updated whenever this action is recieved, but this is cauing black
screen flashes because the active apps are having their overlay paths
reconfigured.

This change makes it so only overlays becoming enabled or disabled cause
the overlay paths to be updated.

Bug: 78508043
Test: Tested Instagram, adding a fingerprint, and adding a new user and
observed no screen flashes.

Change-Id: I3168f4a1e7187c8c1659fb169bae841783b7279f