Fix Notification panel flickers comes from ag/4008279
The flickr is come from the quick setting panel appear. The root
cause is ag/4008279. Because FingerprintUnlockController.showBouncer
will be called after fingerprint authentication, to show the bouncer
without considerating the mode is the problem.

The solution is to use FingerprintUnlockController.calculateMode()
to get whether the mode is MODE_SHOW_BOUNCER or not. The program call
mStatusBarKeyguardViewManager.showBouncer in
FingerprintUnlockController.showBouncer when the mode is
MODE_SHOW_BOUNCER.

Bug: 79614864
Test: atest SystemUITests
Change-Id: I9dc031628d51eb80a05e16f5520c75a74b81d9dc
Fix: 79614864