Go back to using PreferenceThemeOverlay.v14.Material
We temporarily need to go back to having
PreferenceThemeOverlay.SettingsBase inherit from
PreferenceThemeOverlay.v14.Material. This v14.Material version has been
deprecated in the support library source in favor of just having those
styles be what's in PreferenceThemeOverlay, but the prebuilt with those
changes hasn't made it to the master branch yet.

Bug: 110048976
Test: visual (settings should not have dividers between each Preference)
Change-Id: Id9c33f6f235b1cb39cf78830da901413f7aa6815