Ensure dropbox could copy tombstone files to special folder.
Process of tombstoned have changed the generation flow for
tombstone, and dropbox could not be notified when generating new
tombstone file any more, so dropbox could not copy and compress
tombstone file to /data/system/dropbox. We need to modify
observer events from CLOSE_WRITE to CREATE, and it could
work normally.

Bug: http://b/111608961
Test: 1 After tombstone is triggered, we could see the tombstone
      file in the data/system/dropbox directory.

Signed-off-by: Haoran Li <lihaoran5@huawei.com>

Change-Id: I9d6a31773e4a58658ffab214b1e337f27e9f3ae6